+++ sandbox/graph_loop_test.py
@@ -16,26 +16,22 @@
 
 #============= enthought library imports =======================
 from traits.etsconfig.etsconfig import ETSConfig
+
 ETSConfig.toolkit = 'qt4'
 
 from pychron.graph.graph import Graph
-from chaco.plot import Plot
 
-from chaco.lineplot import LinePlot
-from chaco.scatterplot import ScatterPlot
-from pychron.graph.regression_graph import StackedRegressionGraph
 from pychron.core.ui.gui import invoke_in_main_thread
 # from pychron.core.ui.thread import Thread
 from threading import Event, Timer, Thread
 # from pychron.graph.stacked_graph import StackedGraph
 from traits.api import HasTraits, Instance, Button
-from traitsui.api import View, Item, UItem
+from traitsui.api import View, UItem
 # from pychron.consumer_mixin import consumable
 # from pychron.graph.graph import Graph
 # from pychron.graph.regression_graph import RegressionGraph, StackedRegressionGraph
 import time
-from pyface.timer.do_later import do_later, do_after
-from pychron.core.codetools.memory_usage import mem_log, get_current_mem, count_instances
+from pychron.core.codetools.memory_usage import mem_log, get_current_mem
 #============= standard library imports ========================
 #============= local library imports  ==========================