+++ CHANGELOG.md
@@ -33,3 +33,4 @@ CHANGELOG
 * added note, last_modified to irrad_ProductionTable
 * updated editing default_fits
 * added snapshot view to recall
+* added fontsize control to recall